Generic & Reusable Get API Calls with URLSession

+++ WalkAround/Modules/ProductsView/Model/ProductDetailsUIModel.swift
@@ -10,7 +10,7 @@ import Foundation
 struct ProductDetailsUIModel: Decodable {
     let id: Int
     let title: String
-    let price: Float // Not Double as decimal value will not so long
+    let price: Double // Not Double as decimal value will not so long
     let description: String
     let category: String
     let image: String
@@ -18,6 +18,6 @@ struct ProductDetailsUIModel: Decodable {
 }
 
 struct RatingModel: Decodable {
-    let rate: Float
+    let rate: Double
     let count: Int
 }
